git-svn-id: https://svn.kuali.org/repos/student/trunk@14359 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ks-core/ks-common-ui/src/main/java/org/kuali/student/common/ui/client/configurable/mvc/binding/HasValueBinding.java
@@ -1,10 +1,18 @@
-/*
- * Copyright 2009 The Kuali Foundation Licensed under the Educational Community License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License. You may obtain a copy of the License at
- * http://www.osedu.org/licenses/ECL-2.0 Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
- * implied. See the License for the specific language governing permissions and limitations under the License.
+/**
+ * Copyright 2010 The Kuali Foundation Licensed under the
+ * Educational Community License, Version 2.0 (the "License"); you may
+ * not use this file except in compliance with the License. You may
+ * obtain a copy of the License at
+ *
+ * http://www.osedu.org/licenses/ECL-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an "AS IS"
+ * BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
  */
+
 package org.kuali.student.common.ui.client.configurable.mvc.binding;
 
 import java.util.Date;
@@ -12,7 +20,6 @@ import java.util.Date;
 import org.kuali.student.common.ui.client.mvc.DataModel;
 import org.kuali.student.core.assembly.data.Data;
 import org.kuali.student.core.assembly.data.QueryPath;
-import org.kuali.student.core.assembly.data.Data.DataValue;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.user.client.ui.HasValue;
@@ -44,7 +51,7 @@ public class HasValueBinding extends ModelWidgetBindingSupport<HasValue> {
         } else if (value instanceof Double) {
             model.set(qPath, (Double) value);
         } else if (value instanceof Byte) {
-            model.set(qPath, (Integer) value);
+            model.set(qPath, ((Byte) value).intValue());
         } else if (value instanceof Boolean) {
             model.set(qPath, (Boolean) value);
         } else if (value instanceof Date) {
